feat/#62 HomeView, OrderFormView, ColorCell, ConceptView UI
